refactor: 좋아요 테이블 위치 변경
